added all authintications More Authintications Routers and Purchase Routers and Validations fro Deosits

+++ app/routers/products.py
@@ -2,6 +2,7 @@ from fastapi import FastAPI, APIRouter,status,Depends,Response,HTTPException
 from ..util import get_db_conn_cur
 import logging
 from app import schemas
+from app.oauth2 import get_current_user
 
 
 logging.basicConfig(level=logging.ERROR,format='----->%(asctime)s - %(name)s - %(message)s ') 
@@ -33,14 +34,17 @@ async def read_root(Dependcies:tuple=Depends(get_db_conn_cur)):
 
 
 @productRouter.post("/product",status_code=status.HTTP_201_CREATED)
-def create_a_product(product:schemas.productCreate,response:Response,Dependcies:tuple=Depends(get_db_conn_cur)):
+def create_a_product(product:schemas.productCreate,response:Response,Dependcies:tuple=Depends(get_db_conn_cur),current_user:dict=Depends(get_current_user)):
     """This is a POST request to create a new  product."""
     # logging.error(f'Data rom user:Payload received {product.model_dump_json()}-{response.status_code}')
+    logging.error(f"----------------=================CurrentUser=============== {current_user}")
     try:
+        if current_user['role']!='seller':
+            raise HTTPException(status_code=403,detail="Only Sellers can Create Products ")
         conn=Dependcies[0]
         cur=Dependcies[1]
-        cur.execute("""INSERT INTO products (name, price, inventory) VALUES (%s, %s, %s) RETURNING name , price,inventory;""",
-                   (product.name, product.price, product.inventory))
+        cur.execute("""INSERT INTO products (name, price, inventory,seller_id) VALUES (%s, %s, %s,%s) RETURNING *;""",
+            (product.name, product.price, product.inventory,current_user['id']))
         product=cur.fetchone()
         conn.commit()
         logging.error(f'Data received  From DataBase: {product}-{response.status_code}')
@@ -83,16 +87,33 @@ async def update_product(id: int, product: schemas.productUpdate, response: Resp
 
 
 @productRouter.delete("/product/{id}",status_code=status.HTTP_202_ACCEPTED)
-async def update_product(id : int ,response: Response,Dependcies:tuple=Depends(get_db_conn_cur)):
-    try :
-        conn=Dependcies[0]
+async def update_product(id : int ,response: Response,Dependcies:tuple=Depends(get_db_conn_cur),current_user:dict=Depends(get_current_user)):
+    logging.error(f"----------------=================CurrentUser=============== {current_user}")
+    try:
         cur=Dependcies[1]
-        cur.execute("""DELETE from products  where id =%s returning*; """,(id,))
-        update_product=cur.fetchone()
-        conn.commit()
-        logging.debug(f" Updated--{update_product}")
-        return {"Message":update_product}
+        conn=Dependcies[0]
+        cur.execute("""SELECT * from products  where id =%s ;""",(id,))
+        seller_id=int(dict(cur.fetchone())['seller_id'])
+        logging.error(f" Get Seller_id --{seller_id}")
     except Exception as e:
-        logging.error(f"Error getting data from database {e}")
+        logging.error(f"Error getting product info for delete : {e}")
+        return {"message":f"Error {e}"}
+        # seller_id=dict(product)['seller_id']
+
+    try :
+        if current_user['role']!='seller' :
+            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,detail="Only Sellers can Create Products ")
+        if current_user['id'] != seller_id:
+            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,detail="Only Sellers can Create Products ")
+        cur.execute("""DELETE from products  where id =%s AND seller_id =%s returning*; """,(id,current_user['id']))
+        deleted_product=cur.fetchone()
+        conn.commit()
+        logging.error(f" Deleted Product--{deleted_product}")
+        return {"Message":deleted_product}
+    except Exception as e:
+        logging.error(f"Error Deleting data from database {e}")
         conn.rollback()
-        raise HTTPException(status_code=503,detail="Service Unavailable")
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,detail="Unauthorized to perform this action")
+
+
+
